# Conflicts:
#	.gitignore
#	src/main/java/io/github/mzmine/datamodel/features/types/annotations/CompoundDatabaseMatchesType.java
#	src/main/java/io/github/mzmine/modules/dataprocessing/id_localcsvsearch/LocalCSVDatabaseSearchParameters.java
#	src/main/java/io/github/mzmine/modules/dataprocessing/id_localcsvsearch/LocalCSVDatabaseSearchTask.java
#	src/main/java/io/github/mzmine/modules/io/import_rawdata_mzml/ConversionUtils.java
#	src/main/java/io/github/mzmine/util/MirrorChartFactory.java

+++ src/main/java/io/github/mzmine/modules/io/projectsave/ProjectSavingTask.java
@@ -1,19 +1,26 @@
 /*
- * Copyright 2006-2021 The MZmine Development Team
+ * Copyright (c) 2004-2022 The MZmine Development Team
  *
- * This file is part of MZmine.
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
  *
- * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
  *
- * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with MZmine; if not,
- * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
  */
 
 package io.github.mzmine.modules.io.projectsave;
@@ -29,6 +36,7 @@ import io.github.mzmine.taskcontrol.AbstractTask;
 import io.github.mzmine.taskcontrol.TaskStatus;
 import io.github.mzmine.util.ExceptionUtils;
 import io.github.mzmine.util.StreamCopy;
+import io.github.mzmine.util.files.FileAndPathUtil;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -63,7 +71,8 @@ public class ProjectSavingTask extends AbstractTask {
   private UserParameterSaveHandler userParameterSaveHandler;
 
   private final int totalSaveItems;
-  private int currentStage, finishedSaveItems = 0;
+  private final int finishedSaveItems = 0;
+  private int currentStage;
   private String currentSavedObjectName;
 
   // This hashtable maps raw data files to their ID within the saved project
@@ -152,6 +161,14 @@ public class ProjectSavingTask extends AbstractTask {
       // Prepare a temporary ZIP file. We create this file in the same
       // directory as the final saveFile to avoid moving between
       // filesystems in the last stage (renameTo)
+
+      // checking if directory exists first
+      if (!FileAndPathUtil.createDirectory(saveFile.getParentFile())) {
+        setErrorMessage("Could not create directories for file " + saveFile + " for writing.");
+        setStatus(TaskStatus.ERROR);
+        return;
+      }
+
       File tempFile = File.createTempFile(saveFile.getName(), ".tmp", saveFile.getParentFile());
       tempFile.deleteOnExit();
